Refactored an_packets out of the same ugly scripts
an_packets.py: the default script to help reproduce failures, semi-automatic
an_packets-hack.py: very ugly hacks for specific cases when it's hard, due to
the SUT, to find out, [semi-]automatically, if it crashed or not. This can be
due to a TC which caused a failure A during online fuzzing, then causes failure
B during offline failure reproduction (HACK-1); something like HACK-1 with
the aggravating of an SUT which takes time to output logs (HACK-2); or
something like HACK-1 with the aggravating of an SUT which needs gdb for
stacktraces.

This is still ugly and needs to be refactored and rethought. It was just
enough to, semi-automatically, help me with gathering the correct data for
the MSc. Dissertation while in a hurry.